Split Main.hs again, in preparation for more input formats
