+++ languages/haskell/app/Main.hs
@@ -1,50 +1,35 @@
 {-# LANGUAGE NoMonomorphismRestriction, MultiWayIf, OverloadedStrings, DuplicateRecordFields #-}
+{-# OPTIONS_GHC -Wall #-}
 
-module Main
+module Main (main)
 where
 
 import System.IO
-import System.FilePath
 import Control.Monad
 import Data.List.Split (splitOn)
-import Data.List (takeWhile, isSuffixOf, intercalate, nub)
-import Data.Maybe
+import Data.List (intercalate, nub)
 
 import System.Console.Haskeline
 -- import Debug.Trace
 
 import DMN.Types
-import DMN.ParseTable
 import DMN.DecisionTable
 import DMN.Translate.JS
-import DMN.ParseFEEL
 
-import Text.Megaparsec hiding (label)
-import Text.Megaparsec.Char
-import DMN.ParsingUtils
-import qualified Data.Text as T
-import qualified Options.Applicative as OA
-import Options.Applicative (long, short, help, helper, fullDesc, progDesc, strOption, switch, value, info, metavar, str, argument)
 import System.Posix.Terminal (queryTerminal)
 import System.Posix.IO (stdOutput)
 
 import Options
+import ParseMarkdown (parseMarkdown)
 
 
 main :: IO ()
 main = do
   opts <- parseOptions
   myouthandle <- myOutHandle $ out opts
-  let infiles = if null (input opts) then ["-"] else input opts
-  mydtchunks <- mapM (fileChunks opts) (zip [1..] infiles)
-  mydtables1 <- mapM (\mychunk ->
-                           either
-                           (\myPTfail -> do myerr opts $ "** failed to parse table " ++ chunkName mychunk ++ "   :ERROR:\nat " ++ myPTfail
-                                            return Nothing)
-                           (return . Just)
-                           (parseOnly (parseTable (chunkName mychunk) <?> "parseTable") $ T.pack $ unlines $ chunkLines mychunk)
-                       ) (concat mydtchunks)
-  let mydtables = catMaybes mydtables1 -- all this really wants to be an ArrowList
+  -- let infiles = if null (input opts) then ["-"] else input opts
+
+  mydtables <- parseMarkdown opts -- TODO: Don't hardcode markdown here
   mylog opts $ "* imported " ++ show (length mydtables) ++ " tables."
   mylog opts $ "pick = " ++ pick opts
   let pickedTables = if not (null (pick opts)) then filter (\dt -> tableName dt `elem` (trim <$> splitOn "," (pick opts))) mydtables else mydtables
@@ -97,16 +82,7 @@ main = do
 
     -- TODO: validation via incompleteness and conflict detection, under different hit policies
 
-    fileChunks :: ArgOptions -> (Int, FilePath) -> IO InputChunks
-    fileChunks opts (inum,infile) = do
-      mylog opts $ "* opening file: " ++ infile
-      inlines <- if infile == "-" then getContents else readFile infile
-      let rawchunksEither = parseOnly (grepMarkdown ("f"++show inum) <?> "grepMarkdown") (T.pack inlines)
-      either
-        (\errstr -> myerr opts ("** parser failure in grepMarkdown: " ++ errstr) >> return [])
-        return
-        rawchunksEither
-    myerr opts = hPutStrLn stderr
+    myerr _ = hPutStrLn stderr
     mylog opts msg = when (verbose opts) $ myerr opts msg
 
     differentlyTyped :: [DecisionTable] -> Bool
@@ -117,7 +93,7 @@ main = do
 
 -- not quite finished; in future refactor this over to JS.hs
 showToJSON :: DecisionTable -> [[FEELexp]] -> [String]
-showToJSON dtable cols = if not (null cols) then zipWith showFeels ((getOutputHeaders . header) dtable) cols else []
+showToJSON dtable cols' = if not (null cols') then zipWith showFeels ((getOutputHeaders . header) dtable) cols' else []
 
 outputTo :: Handle -> String -> ArgOptions -> DecisionTable -> IO ()
 outputTo h "js" opts dtable = hPutStrLn h $ toJS (JSOpts (Options.propstyle opts) (outformat opts == "ts")) dtable
@@ -127,59 +103,3 @@ myOutHandle :: FilePath -> IO Handle
 myOutHandle h = if h == "-" then return stdout else openFile h WriteMode
 
 --  putStrLn $ toJS (fromRight (error "parse error") (parseOnly (parseTable "mydmn1") dmn2))
-
-type InputChunks = [InputChunk]
-
-data InputChunk = InputChunk
-  { chunkName  :: String
-  , chunkLines :: [String]
-  }
-  deriving (Show, Eq)
--- in future, consider grabbing the tables out of Pandoc -- maybe this would be better off as a JSON filter?
-
-grepMarkdown :: String -> Parser InputChunks
-grepMarkdown defaultName = do
-  mytables <- many1 (try (grepTable defaultName) <?> "grepTable")
-  (many irrelevantLine >> eof >> return Nothing)
-  return $ catMaybes mytables
-
-grepTable :: String -> Parser (Maybe InputChunk)
-grepTable defaultName = do
-  mHeader <- maybeHeaderLines <?> "maybeHeaderLines"
-  tablelines <- many1 (getTableLine <?> "getTableLine")
-  return (Just (InputChunk (fromMaybe defaultName mHeader) tablelines))
-
-getTableLine :: Parser String
-getTableLine = do
-  pipe <- char '|'
-  therest <- manyTill anyChar endOfLine
-  return $ pipe : therest
-
-irrelevantLine :: Parser (Maybe String)
-irrelevantLine = do
-  endOfLine <|> (notChar '|' >> skipWhile "character" (/= '\n') >> endOfLine)
-  return Nothing
-
-maybeHeaderLines :: Parser (Maybe String)
-maybeHeaderLines = do
-  gotHeaders <- catMaybes <$> many maybeHeaderLine
-  return (if not (null gotHeaders) then Just (last gotHeaders) else Nothing)
-
-maybeHeaderLine :: Parser (Maybe String)
-maybeHeaderLine = do
-  foundLine <- try orgNameLine <|> try (headerLine <?> "header line") <|> (irrelevantLine <?> "irrelevant line")
-  return $ Data.List.takeWhile (/=':') <$> foundLine
-
-headerLine :: Parser (Maybe String)
-headerLine = do
-  skipMany1 (satisfy (\c -> c == '#' || c == '*'))
-  skipHorizontalSpace
-  content <- manyTill anyChar endOfLine
-  return (Just content)
-
-orgNameLine :: Parser (Maybe String)
-orgNameLine = do
-  "#+NAME:" >> skipHorizontalSpace
-  content <- manyTill anyChar endOfLine
-  return (Just content)
-
