+++ languages/haskell/dmnmd.cabal
@@ -67,6 +67,7 @@ executable dmnmd
   main-is: Main.hs
   other-modules:
       Options
+      ParseMarkdown
       Paths_dmnmd
   hs-source-dirs:
       app
